requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib.pyplot==3.8.2
yfinance==0.2.33
requests==2.31.0
